remove pkgsign tools
