+++ conan/internal/rest/pkg_sign.py
@@ -1,13 +1,97 @@
 import os
+import json
 
+from conan.api.output import ConanOutput
 from conan.errors import ConanException
 from conan.internal.cache.conan_reference_layout import METADATA
 from conan.internal.cache.home_paths import HomePaths
 from conan.internal.loader import load_python_file
-from conan.internal.util.files import mkdir
-from conan.tools.pkg_signing.plugin import (get_manifest_filepath, get_signatures_filepath,
-                                            _save_manifest, _save_signatures, PKGSIGN_MANIFEST,
-                                            PKGSIGN_SIGNATURES, _verify_files_checksums)
+from conan.internal.util.files import load, mkdir, save, sha256sum
+
+
+PKGSIGN_MANIFEST = "pkgsign-manifest.json"
+PKGSIGN_SIGNATURES = "pkgsign-signatures.json"
+
+
+def _save_manifest(artifacts_folder, signature_folder):
+    """
+    Creates the summary content as a dictionary for manipulation.
+
+    Returns a structure like:
+        {
+            "files": [
+                {"file": "conan_package.tgz", "sha256": "abc123"},
+                {"file": "other_file.bin", "sha256": "fff999"},
+                ...
+            ]
+        }
+    """
+    files_list = []
+
+    for fname in os.listdir(artifacts_folder):
+        file_path = os.path.join(artifacts_folder, fname)
+
+        if os.path.isfile(file_path):
+            entry = {
+                "file": fname,
+                "sha256": sha256sum(file_path)
+            }
+            files_list.append(entry)
+
+    # Sort files by filename to ensure consistent order
+    files_list.sort(key=lambda x: x["file"])
+    save(os.path.join(signature_folder, PKGSIGN_MANIFEST),
+         json.dumps({"files": files_list}, indent=2))
+
+
+def _save_signatures(signature_folder, signatures):
+    """
+    Saves the content of signatures file in the signature folder
+    :param signature_folder: Signature folder path
+    :param signatures: dict of {filename: signature_value}
+    """
+    for signature in signatures:
+        assert signature.get("method"), "Signature 'method' must be set"
+        assert signature.get("provider"), "Signature 'provider' must be set"
+        assert signature.get("sign_artifacts"), "Signature 'sign_artifacts' must be set"
+        assert isinstance(signature.get("sign_artifacts"), dict), \
+            "'sign_artifacts' must be a dict of {name: signature_filename}"
+    assert os.path.isfile(os.path.join(signature_folder, PKGSIGN_MANIFEST)),\
+        "Manifest file must exist before saving signatures"
+    content = {
+        "manifest": PKGSIGN_MANIFEST,
+        "signatures": signatures
+    }
+    save(os.path.join(signature_folder, PKGSIGN_SIGNATURES), json.dumps(content, indent=2))
+
+
+def _verify_files_checksums(signature_folder, files):
+    """
+    Verifies that the files' checksums match those stored in the summary.
+    :param signature_folder: Signature folder path
+    :param files: dict of {filename: filepath} of files in artifact folder to verify
+    """
+    if not os.path.isfile(os.path.join(signature_folder, PKGSIGN_MANIFEST)):
+        raise ConanException(f"Manifest file does not exist in signature folder "
+                             f"{os.path.join(signature_folder, PKGSIGN_MANIFEST)}")
+
+    manifest_content = load(os.path.join(signature_folder, PKGSIGN_MANIFEST))
+    expected_list = json.loads(manifest_content).get("files", [])
+    expected_files = {item["file"]: item["sha256"] for item in expected_list}
+
+    # This is checking that the files of the package exist in the manifest instead of the opposite
+    # because some files might be missing such as conan_sources.tgz
+    for filename, file_path in files.items():
+        expected_checksum = expected_files.get(filename)
+        actual_checksum = sha256sum(file_path)
+
+        if actual_checksum != expected_checksum:
+            raise ConanException(
+                f"Checksum mismatch for file {filename}: "
+                f"expected {expected_checksum}, got {actual_checksum}."
+            )
+        else:
+            ConanOutput().info(f"Checksum verified for file {filename} ({actual_checksum}).")
 
 
 class PkgSignaturesPlugin:
@@ -42,8 +126,8 @@ class PkgSignaturesPlugin:
                                  "a list of signature dicts")
         _save_signatures(metadata_sign, signatures)
         # Add files to package bundle so they get uploaded
-        files[f"{METADATA}/sign/{PKGSIGN_MANIFEST}"] = get_manifest_filepath(metadata_sign)
-        files[f"{METADATA}/sign/{PKGSIGN_SIGNATURES}"] = get_signatures_filepath(metadata_sign)
+        files[f"{METADATA}/sign/{PKGSIGN_MANIFEST}"] = os.path.join(metadata_sign, PKGSIGN_MANIFEST)
+        files[f"{METADATA}/sign/{PKGSIGN_SIGNATURES}"] = os.path.join(metadata_sign, PKGSIGN_SIGNATURES)
         for sig in signatures:
             for name, file in sig.get("sign_artifacts", {}).items():
                 #TODO: print output?
